feat(otter-go): regenerate db/generated with sqlc and gate drift in CI (#384)
* feat(otter-go): regenerate db/generated with sqlc and gate drift in CI

Revives the sqlc drift gate that was retired with the alembic rip:
CI now migrates a throwaway Timescale Postgres, dumps the schema,
regenerates db/generated with pinned sqlc v1.31.1, and fails on any
diff. To make that possible, db/generated is regenerated for real for
the first time since the files went hand-maintained — sqlc's output
is now byte-for-byte the committed code.

Fixing the sources so regeneration reproduces the hand-written API
surfaced four latent runtime bugs the drift gate now prevents:

* Audit writes have been silently failing: InsertAuditLog targeted
  "audit_logs" but the table is "audit_log" (the recorder logs and
  continues by design, so every audited mutation wrote nothing —
  0 rows in the dev DB).
* BGP mutations cast to ::asn_kind / ::community_kind /
  ::policy_action, none of which exist — the real enum types are
  bgp_-prefixed. CreateAsn/UpdateAsn and friends 500'd at runtime.
* CreateIPAddress/UpdateIPAddress cast to ::ip_address_role/
  status/source; the real types are ip_role/ip_status/ip_source.
* ApproveLirRequest documented a $3 parameter that was never bound
  (mixed positional/named args left a numbering hole).

Main moving parts:

* db/sqlc.yaml: full override set pinning the wire conventions the
  handlers were written against — enums as plain string, uuid/
  timestamptz/date as stdlib types, NUMERIC as decimal string,
  json/jsonb as json.RawMessage (encoding/json base64-encodes
  []byte; RawMessage keeps raw JSON for rows marshaled straight
  into responses), inet/cidr/macaddr as text. Rename map restores
  the hand initialisms (IPFamily, TTL, KeaURL, PoolsJSON, ...).
  dns_keys.private_pem keeps its json:"-" tag via a column
  override so DNSSEC private keys can never serialize.
* db/queries: cast-wrapped positional params converted to named
  sqlc.arg()/sqlc.narg() so Params field names survive
  regeneration; full-projection queries collapsed to
  SELECT */RETURNING * so they return model structs again;
  computed inet/cidr projections wrapped ::text (pgx cannot scan
  binary inet/cidr into strings); nullable host() projections
  CASE-wrapped to keep NULL representable; :execrows restored
  where callers read RowsAffected.
* Queries that project secrets keep deliberate per-query rows:
  collector rows exclude enrollment_token_hash (the regenerated
  Collector model now carries it, and the list/get handlers
  marshal rows directly), admin user rows exclude password_hash.
* Callers updated for honest types the regen exposed: non-null
  columns dropped pointer wrappers, max_amps/cord_length_m convert
  int/float to the wire's decimal strings at the boundary, the LIR
  approval handler unpacks the CTE's flat row back into the
  request/allocation response shape.
* ci.yml: sqlc-drift job (Timescale service container, migrate,
  filtered pg_dump, sqlc generate, git diff --exit-code).

go build/vet clean; all 49 test packages pass, including the
models_jsonb_test wire guard.

* build: align go.work and module toolchains on go 1.26.0

The otter-go go.mod toolchain bump rode along with the sqlc
regeneration commit while go.work still declared 1.25.7, which
breaks every workspace member's build. Commit the rest of the
already-local toolchain drift so the workspace is consistent.

* fix(otter-go): repair runtime breaks the compiler could not see

Reproduced the PR-CI E2E failures against a scratch database and
fixed three runtime-only regressions from the regeneration:

* GetUserScopedCapabilities: the scope_type::text cast erased the
  LEFT JOIN nullability sqlc models correctly when uncast — a NULL
  scope_type (any unrestricted role assignment) failed the scan.
  Dropped the cast; the field is *string again as the resolver
  expects.
* ListAssetsByRackOrdered / ListSiteAlertsBySeverity /
  ListAssetIPAddresses / ListRecentAssetAlerts: filters on nullable
  columns made sqlc emit *uuid.UUID params where the dashboards'
  narrow Querier interfaces expect uuid.UUID — and those interfaces
  are asserted at RUNTIME (h.Q.(RackDetailQuerier)), so the mismatch
  surfaced as 'rack-detail requires full Querier' 500s, not compile
  errors. Forced non-null with sqlc.arg(...)::uuid.
* Added querier_guards.go: compile-time
  var _ XxxQuerier = (*dbq.Queries)(nil) assertions for all seven
  dashboards narrow interfaces so this class can never be silent
  again.

Verified against a migrated scratch DB: login, rack/site/building
detail, forecast, free-space, collectors, asset create — and the
first real audit_log rows this codebase has ever written.

* ci: pin sqlc via go.mod tool directive; fix Sonar gate config

* sqlc now runs as 'go tool sqlc' — version and hashes pinned by
  go.mod/go.sum, replacing the go-install step Sonar flagged as an
  unpredictable dependency. Verified byte-identical generation.
* sonar-project.properties: collapse the exclusions to one line
  (generated-code issues were leaking through the continuation
  form) and exclude db/** from copy-paste detection — per-enum
  override blocks and per-table CRUD queries are repetitive by
  nature and were tripping the new-code duplication gate.

* ci: configure SonarCloud automatic analysis via .sonarcloud.properties

Automatic analysis reads .sonarcloud.properties from the analyzed
branch — NOT sonar-project.properties, which is for CI-based
scanners and has demonstrably never applied here (main still
carries indexed measures for db/generated and the migrations dir
despite the exclusions declared there since onboarding). This is
why PR 384's regenerated db/generated tripped the new-code
duplication gate: the files were never actually excluded.

Mirror the intended analysis scope into the file automatic
analysis actually honors, and mark sonar-project.properties as
CI-scanner-only.

+++ packages/otter-go/db/queries/dhcp_push.sql
@@ -56,7 +56,9 @@ WHERE id = $1;
 -- the lowest free positive int. A production fleet might persist a
 -- per-server sequence; the O(n) scan matches Python's posture and
 -- is fine until a server has thousands of scopes (a wild outlier).
-SELECT kea_subnet_id
+-- The ::int cast makes the WHERE-guaranteed non-NULL visible to
+-- sqlc (the column itself is nullable).
+SELECT kea_subnet_id::int
 FROM dhcp_scopes
 WHERE dhcp_server_id = $1
   AND kea_subnet_id IS NOT NULL
